+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 - Database populatation now works only for Odoo version 14-17, because starting from Odoo 18 population means duplication instead of generation.
 - After backup of database completed, log message about backup completed and duration of operation.
 
+### Fixed
+
+- Fixed bug with false-positives in when running migration tests on new addons.
+
 ---
 
 ## Release 0.5.1 (2025-11-02)